feat: 로그인 버튼의 이동 및 로그아웃 기능 구현
